:fire: Simplify depth datamodules
